+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - **chai** has been upgraded from 4.2.0 to **4.3.4**
 - **grunt-karma** has been removed
+- **grunt-mkdocs** has been removed
 - **karma** has been upgraded from 4.4.1 to **6.3.4**
 - **karma-mocha** has been upgraded from 1.3.0 to **2.0.1**
 - **mocha** has been upgraded from 7.1.1 to **9.1.1**